Make different files for board and player logic
