+++ lib/player.rb
@@ -1,0 +1,7 @@
+class Player
+    attr_reader :name, :letter
+    def initialize(name, letter)
+      @name = name
+      @letter = letter
+    end
+  end
